mcp-servers/threads-mcp-server/pyproject.toml
[project]
name = "threads-mcp-server"
version = "0.1.0"
description = "Meta Threads MCP server for posting and managing Threads content"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["*.py"]
